+++ packages/clerk-js/src/ui/components/UserProfile/UserProfileRoutes.tsx
@@ -29,7 +29,7 @@ export const UserProfileRoutes = () => {
   const isAccountPageRoot = pages.routes[0].id === USER_PROFILE_NAVBAR_ROUTE_ID.ACCOUNT;
   const isSecurityPageRoot = pages.routes[0].id === USER_PROFILE_NAVBAR_ROUTE_ID.SECURITY;
   const isBillingPageRoot = pages.routes[0].id === USER_PROFILE_NAVBAR_ROUTE_ID.BILLING;
-  const isApiKeysPageRoot = pages.routes[0].id === USER_PROFILE_NAVBAR_ROUTE_ID.API_KEYS;
+  const isAPIKeysPageRoot = pages.routes[0].id === USER_PROFILE_NAVBAR_ROUTE_ID.API_KEYS;
 
   const customPageRoutesWithContents = pages.contents?.map((customPage, index) => {
     const shouldFirstCustomItemBeOnRoot = !isAccountPageRoot && !isSecurityPageRoot && index === 0;
@@ -95,7 +95,7 @@ export const UserProfileRoutes = () => {
           </Route>
         )}
         {apiKeysSettings.enabled && (
-          <Route path={isApiKeysPageRoot ? undefined : 'api-keys'}>
+          <Route path={isAPIKeysPageRoot ? undefined : 'api-keys'}>
             <Switch>
               <Route index>
                 <Suspense fallback={''}>